+++ Assets/Scripts/GUI/ImportCreature.cs
@@ -84,7 +84,7 @@ public class ImportCreature : MonoBehaviour
                 float bjf = float.Parse(contents["attributes"]["base_joint_frequency"].ToString());
                 float bja = float.Parse(contents["attributes"]["base_joint_amplitude"].ToString());
                 float bjp = float.Parse(contents["attributes"]["base_joint_phase"].ToString());
-                float ht = float.Parse(contents["attributes"]["hunger_threshold"].ToString());
+                decimal ht = decimal.Parse(contents["attributes"]["hunger_threshold"].ToString());
 
                 ArrayList branches = new ArrayList();
                 int num_branches = (int)contents["attributes"]["branches"];
